requirements.txt
annotated-types==0.7.0
anyio==4.4.0
fastapi==0.112.2
greenlet==3.0.3
idna==3.8
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.33
starlette==0.38.4
typing_extensions==4.12.2
